requirements.txt
groq
requests
Pillow
pyautogui
pyperclip
